update oci-go-sdk to 2.0.0 for lb tag support (#219)

+++ pkg/oci/load_balancer_spec.go
@@ -314,7 +314,7 @@ func getListeners(svc *v1.Service, sslCfg *SSLConfig) (map[string]loadbalancer.L
 
 		if connectionIdleTimeout > 0 {
 			listener.ConnectionConfiguration = &loadbalancer.ConnectionConfiguration{
-				IdleTimeout: common.Int(int(connectionIdleTimeout)),
+				IdleTimeout: common.Int64(int64(connectionIdleTimeout)),
 			}
 		}
